@@ -1,66 +1,66 @@
 name: mindflow_ai_coach
-description: The first AI coach that actually understands how YOUR brain works. Uses NLP psychology frameworks for personalized coaching.
+description: The first AI coach that actually understands how YOUR brain works.
 publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.2.0 <4.0.0'
+  sdk: '>=3.3.0 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  
-  # Firebase
-  firebase_core: ^3.1.0
-  firebase_auth: ^5.1.0
-  cloud_firestore: ^5.0.0
-  
-  # Google Sign-In
+
+  # --- Core Firebase ---
+  firebase_core: ^3.4.0
+  firebase_auth: ^5.1.4
+  cloud_firestore: ^5.4.0
+
+  # --- Authentication ---
   google_sign_in: ^6.2.1
-  
-  # AI - Gemini
-  google_generative_ai: ^0.2.2
-  
-  # Monetization - RevenueCat
-  purchases_flutter: ^6.29.0
-  
-  # State Management
+
+  # --- AI (Gemini) ---
+  google_generative_ai: ^0.4.0
+
+  # --- Money (RevenueCat) ---
+  purchases_flutter: ^8.0.0
+  purchases_ui_flutter: ^8.0.0
+
+  # --- State Management (Using Riverpod as primary, Provider kept for legacy) ---
+  flutter_riverpod: ^2.6.1
   provider: ^6.1.2
-  
-  # Navigation
-  go_router: ^14.1.0
-  
-  # UI Enhancements
-  lottie: ^3.1.2  # Animations
-  
-  # Utilities
-  intl: ^0.19.0  # Date formatting
-  flutter_dotenv: ^5.1.0  # Environment variables
-  http: ^1.2.0  # HTTP client for backend API calls
-  shared_preferences: ^2.2.2  # Local storage for user data
-  url_launcher: ^6.2.2  # Open Stripe checkout URL
-  uuid: ^4.3.3  # Generate unique user IDs
-  web: ^1.0.0  # JS interop for web platform
-  
-  # Cupertino Icons
+
+  # --- Navigation ---
+  go_router: ^14.2.0
+
+  # --- UI & Media ---
+  lottie: ^3.1.2
   cupertino_icons: ^1.0.6
-  purchases_ui_flutter: ^7.0.0
+  audioplayers: ^6.0.0
+
+  # --- Utilities ---
+  intl: ^0.19.0
+  flutter_dotenv: ^5.1.0
+  http: ^1.2.0
+  shared_preferences: ^2.2.2
+  url_launcher: ^6.2.2
+  uuid: ^4.3.3
+
+  # --- Security ---
+  crypto: ^3.0.3
+  encrypt: ^5.0.3
+  flutter_secure_storage: ^9.0.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   flutter_lints: ^4.0.0
-  mockito: ^5.4.4
   build_runner: ^2.4.9
 
 flutter:
   uses-material-design: true
-  
   assets:
+    - assets/env
     - assets/images/
     - assets/animations/
     - assets/icons/
-  
-  # fonts:
-  #   # Using system fonts (SF Pro on iOS, Roboto on Android)
-  #   # Custom fonts can be added here if needed
+    - assets/audio/
