@@ -44,7 +44,7 @@ class _AmbientSoundPickerState extends State<AmbientSoundPicker> {
       _selectedSound = soundType;
       _showVolumeSlider = soundType != AmbientSoundType.none;
     });
-    
+
     await _soundService.play(soundType);
     widget.onSoundChanged?.call(soundType);
   }
@@ -105,7 +105,7 @@ class _AmbientSoundPickerState extends State<AmbientSoundPicker> {
                 fontFamily: 'DM Sans',
                 fontSize: 14,
                 fontWeight: FontWeight.w600,
-                color: AppColors.jobsObsidian.withOpacity(0.6),
+                color: AppColors.jobsObsidian.withValues(alpha: 0.6),
               ),
             ),
           ),
@@ -117,11 +117,11 @@ class _AmbientSoundPickerState extends State<AmbientSoundPicker> {
             itemBuilder: (context, index) {
               final sound = AmbientSoundService.availableSounds[index];
               final isSelected = _selectedSound == sound.type;
-              
+
               return Padding(
                 padding: EdgeInsets.only(
-                  right: index < AmbientSoundService.availableSounds.length - 1 
-                      ? AppSpacing.spacing12 
+                  right: index < AmbientSoundService.availableSounds.length - 1
+                      ? AppSpacing.spacing12
                       : 0,
                 ),
                 child: GestureDetector(
@@ -130,20 +130,21 @@ class _AmbientSoundPickerState extends State<AmbientSoundPicker> {
                     duration: const Duration(milliseconds: 200),
                     width: 70,
                     decoration: BoxDecoration(
-                      color: isSelected 
-                          ? AppColors.jobsSage.withOpacity(0.2) 
+                      color: isSelected
+                          ? AppColors.jobsSage.withValues(alpha: 0.2)
                           : Colors.white,
                       borderRadius: BorderRadius.circular(16),
                       border: Border.all(
-                        color: isSelected 
-                            ? AppColors.jobsSage 
+                        color: isSelected
+                            ? AppColors.jobsSage
                             : Colors.transparent,
                         width: 2,
                       ),
                       boxShadow: isSelected
                           ? [
                               BoxShadow(
-                                color: AppColors.jobsSage.withOpacity(0.2),
+                                color:
+                                    AppColors.jobsSage.withValues(alpha: 0.2),
                                 blurRadius: 8,
                                 offset: const Offset(0, 4),
                               ),
@@ -163,12 +164,11 @@ class _AmbientSoundPickerState extends State<AmbientSoundPicker> {
                           style: TextStyle(
                             fontFamily: 'DM Sans',
                             fontSize: 11,
-                            fontWeight: isSelected 
-                                ? FontWeight.w600 
-                                : FontWeight.w500,
+                            fontWeight:
+                                isSelected ? FontWeight.w600 : FontWeight.w500,
                             color: isSelected
                                 ? AppColors.jobsSage
-                                : AppColors.jobsObsidian.withOpacity(0.7),
+                                : AppColors.jobsObsidian.withValues(alpha: 0.7),
                           ),
                           textAlign: TextAlign.center,
                           maxLines: 1,
@@ -188,16 +188,17 @@ class _AmbientSoundPickerState extends State<AmbientSoundPicker> {
             children: [
               Icon(
                 Icons.volume_down_rounded,
-                color: AppColors.jobsObsidian.withOpacity(0.5),
+                color: AppColors.jobsObsidian.withValues(alpha: 0.5),
                 size: 20,
               ),
               Expanded(
                 child: SliderTheme(
                   data: SliderThemeData(
                     activeTrackColor: AppColors.jobsSage,
-                    inactiveTrackColor: AppColors.jobsSage.withOpacity(0.2),
+                    inactiveTrackColor:
+                        AppColors.jobsSage.withValues(alpha: 0.2),
                     thumbColor: AppColors.jobsSage,
-                    overlayColor: AppColors.jobsSage.withOpacity(0.1),
+                    overlayColor: AppColors.jobsSage.withValues(alpha: 0.1),
                     trackHeight: 4,
                     thumbShape: const RoundSliderThumbShape(
                       enabledThumbRadius: 8,
@@ -213,7 +214,7 @@ class _AmbientSoundPickerState extends State<AmbientSoundPicker> {
               ),
               Icon(
                 Icons.volume_up_rounded,
-                color: AppColors.jobsObsidian.withOpacity(0.5),
+                color: AppColors.jobsObsidian.withValues(alpha: 0.5),
                 size: 20,
               ),
             ],
@@ -245,14 +246,14 @@ class _TabButton extends StatelessWidget {
         duration: const Duration(milliseconds: 200),
         padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
         decoration: BoxDecoration(
-          color: isSelected 
-              ? AppColors.jobsSage.withOpacity(0.15) 
+          color: isSelected
+              ? AppColors.jobsSage.withValues(alpha: 0.15)
               : Colors.transparent,
           borderRadius: BorderRadius.circular(20),
           border: Border.all(
-            color: isSelected 
-                ? AppColors.jobsSage 
-                : AppColors.jobsObsidian.withOpacity(0.15),
+            color: isSelected
+                ? AppColors.jobsSage
+                : AppColors.jobsObsidian.withValues(alpha: 0.15),
             width: 1,
           ),
         ),
