@@ -1,6 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
+import 'package:provider/provider.dart' as provider;
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:flutter/physics.dart';
 
+import 'core/config/env_config.dart';
 import 'core/theme/jobs_theme.dart';
 import 'core/theme/theme_provider.dart';
 import 'core/constants/app_colors.dart';
@@ -25,64 +28,61 @@ import 'features/profile/presentation/screens/profile_screen.dart';
 import 'features/habits/presentation/screens/habits_screen.dart';
 import 'features/chat/presentation/screens/chat_screen.dart';
 
-void main() {
+Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  runApp(const MindFlowApp());
+  await EnvConfig.initialize();
+  runApp(const ProviderScope(child: MindFlowApp()));
 }
 
-class MindFlowApp extends StatelessWidget {
+class MindFlowApp extends ConsumerWidget {
   const MindFlowApp({super.key});
 
   @override
-  Widget build(BuildContext context) {
-    return MultiProvider(
+  Widget build(BuildContext context, WidgetRef ref) {
+    // Watch the Riverpod theme provider
+    final themeMode = ref.watch(themeProvider);
+
+    return provider.MultiProvider(
       providers: [
-        ChangeNotifierProvider(
-          create: (_) => ThemeProvider(),
-        ),
-        ChangeNotifierProvider(
-          create: (_) => UserProvider()..initialize(),
-        ),
-        ChangeNotifierProvider(
+        // ThemeProvider is now handled by Riverpod
+        // UserProvider is now handled by Riverpod
+        provider.ChangeNotifierProvider(
           create: (_) => ChatProvider()..initialize(),
         ),
-        ChangeNotifierProvider(
+        provider.ChangeNotifierProvider(
           create: (_) => SubscriptionProvider(),
         ),
-        ChangeNotifierProvider(
+        provider.ChangeNotifierProvider(
           create: (_) => HabitProvider()..initialize(),
         ),
-        ChangeNotifierProvider(
+        provider.ChangeNotifierProvider(
           create: (_) => WisdomProvider()..initialize(),
         ),
-        ChangeNotifierProvider(
+        provider.ChangeNotifierProvider(
           create: (_) => BackgroundCoachProvider()..initialize(),
         ),
       ],
-      child: Consumer<ThemeProvider>(
-        builder: (context, themeProvider, child) {
-          return MaterialApp(
-            title: 'MindFlow',
-            debugShowCheckedModeBanner: false,
-            theme: JobsTheme.lightTheme,
-            darkTheme: JobsTheme.darkTheme,
-            themeMode: themeProvider.themeMode,
-            home: const WelcomeScreen(),
-          );
-        },
+      child: MaterialApp(
+        title: 'MindFlow',
+        debugShowCheckedModeBanner: false,
+        theme: JobsTheme.lightTheme,
+        darkTheme: JobsTheme.darkTheme,
+        themeMode: themeMode,
+        home: const WelcomeScreen(),
       ),
     );
   }
 }
 
-class MainAppShell extends StatefulWidget {
+class MainAppShell extends ConsumerStatefulWidget {
   const MainAppShell({super.key});
 
   @override
-  State<MainAppShell> createState() => _MainAppShellState();
+  ConsumerState<MainAppShell> createState() => _MainAppShellState();
 }
 
-class _MainAppShellState extends State<MainAppShell> with WidgetsBindingObserver {
+class _MainAppShellState extends ConsumerState<MainAppShell>
+    with WidgetsBindingObserver {
   int _currentIndex = 0;
 
   final List<Widget> _screens = const [
@@ -97,7 +97,10 @@ class _MainAppShellState extends State<MainAppShell> with WidgetsBindingObserver
   void initState() {
     super.initState();
     WidgetsBinding.instance.addObserver(this);
+
     WidgetsBinding.instance.addPostFrameCallback((_) {
+      // Initialize UserProvider safely after build
+      ref.read(userProvider.notifier).initialize();
       _checkForInterventions();
     });
   }
@@ -117,22 +120,21 @@ class _MainAppShellState extends State<MainAppShell> with WidgetsBindingObserver
   }
 
   void _checkForInterventions() {
-    final userProvider = context.read<UserProvider>();
+    final userState = ref.read(userProvider);
     final habitProvider = context.read<HabitProvider>();
     final coachProvider = context.read<BackgroundCoachProvider>();
 
     final now = DateTime.now();
-    final lastSessionDate = userProvider.sessionsCompleted > 0 ? now : null;
-    final daysSinceLastSession = lastSessionDate != null
-        ? now.difference(lastSessionDate).inDays
-        : 0;
+    final lastSessionDate = userState.sessionsCompleted > 0 ? now : null;
+    final daysSinceLastSession =
+        lastSessionDate != null ? now.difference(lastSessionDate).inDays : 0;
 
     final hasCompletedTodaySession = habitProvider.todayProgress > 0;
 
     coachProvider.onAppResumed(
-      currentStreak: userProvider.currentStreak,
-      totalSessions: userProvider.sessionsCompleted,
-      totalMinutes: userProvider.totalMinutes,
+      currentStreak: userState.currentStreak,
+      totalSessions: userState.sessionsCompleted,
+      totalMinutes: userState.totalMinutes,
       daysSinceLastSession: daysSinceLastSession,
       habits: habitProvider.activeHabits,
       hasCompletedTodaySession: hasCompletedTodaySession,
@@ -145,20 +147,19 @@ class _MainAppShellState extends State<MainAppShell> with WidgetsBindingObserver
       _currentIndex = index;
     });
 
-    final userProvider = context.read<UserProvider>();
+    final userState = ref.read(userProvider);
     final habitProvider = context.read<HabitProvider>();
     final coachProvider = context.read<BackgroundCoachProvider>();
 
     final now = DateTime.now();
-    final lastSessionDate = userProvider.sessionsCompleted > 0 ? now : null;
-    final daysSinceLastSession = lastSessionDate != null
-        ? now.difference(lastSessionDate).inDays
-        : 0;
+    final lastSessionDate = userState.sessionsCompleted > 0 ? now : null;
+    final daysSinceLastSession =
+        lastSessionDate != null ? now.difference(lastSessionDate).inDays : 0;
 
     coachProvider.onTabChanged(
-      currentStreak: userProvider.currentStreak,
-      totalSessions: userProvider.sessionsCompleted,
-      totalMinutes: userProvider.totalMinutes,
+      currentStreak: userState.currentStreak,
+      totalSessions: userState.sessionsCompleted,
+      totalMinutes: userState.totalMinutes,
       daysSinceLastSession: daysSinceLastSession,
       habits: habitProvider.activeHabits,
       hasCompletedTodaySession: habitProvider.todayProgress > 0,
@@ -230,6 +231,7 @@ class _CoachFloatingButtonState extends State<_CoachFloatingButton>
   @override
   void initState() {
     super.initState();
+    // Ambient breathing
     _controller = AnimationController(
       duration: const Duration(milliseconds: 2000),
       vsync: this,
@@ -242,6 +244,10 @@ class _CoachFloatingButtonState extends State<_CoachFloatingButton>
     _glowAnimation = Tween<double>(begin: 0.3, end: 0.5).animate(
       CurvedAnimation(parent: _controller, curve: Curves.easeInOut),
     );
+
+    // Spring press interaction will be handled by a separate widget wrapper
+    // or we can implement it here if we want custom physics.
+    // For now, let's strictly wrap the FAB in a springable widget or custom logic.
   }
 
   @override
@@ -262,24 +268,113 @@ class _CoachFloatingButtonState extends State<_CoachFloatingButton>
               shape: BoxShape.circle,
               boxShadow: [
                 BoxShadow(
-                  color: AppColors.jobsSage.withOpacity(_glowAnimation.value),
+                  color: AppColors.jobsSage
+                      .withValues(alpha: _glowAnimation.value),
                   blurRadius: 20,
                   spreadRadius: 2,
                 ),
               ],
             ),
-            child: FloatingActionButton(
+            // Using a physics-based pressable widget instead of raw FAB
+            child: _SpringPressable(
               onPressed: widget.onPressed,
-              backgroundColor: AppColors.jobsSage,
-              elevation: 8,
-              child: const Text(
-                '🧘',
-                style: TextStyle(fontSize: 24),
+              child: Container(
+                height: 56,
+                width: 56,
+                decoration: const BoxDecoration(
+                  shape: BoxShape.circle,
+                  color: AppColors.jobsSage,
+                ),
+                child: const Icon(Icons.self_improvement_rounded,
+                    color: Colors.white, size: 24),
               ),
             ),
           ),
         );
       },
+    );
+  }
+}
+
+class _SpringPressable extends StatefulWidget {
+  final VoidCallback onPressed;
+  final Widget child;
+
+  const _SpringPressable({required this.onPressed, required this.child});
+
+  @override
+  State<_SpringPressable> createState() => _SpringPressableState();
+}
+
+class _SpringPressableState extends State<_SpringPressable>
+    with SingleTickerProviderStateMixin {
+  late AnimationController _controller;
+
+  @override
+  void initState() {
+    super.initState();
+    _controller = AnimationController(
+      vsync: this,
+      duration: const Duration(milliseconds: 200), // Upper bound
+    );
+  }
+
+  void _onTapDown(TapDownDetails details) {
+    _controller.animateWith(
+      SpringSimulation(
+        const SpringDescription(mass: 1, stiffness: 300, damping: 20),
+        0.0,
+        1.0, // Compressed state
+        0.0,
+      ),
+    );
+  }
+
+  void _onTapUp(TapUpDetails details) {
+    _controller.animateWith(
+      SpringSimulation(
+        const SpringDescription(mass: 1, stiffness: 300, damping: 20),
+        1.0,
+        0.0, // Back to normal
+        0.0,
+      ),
+    );
+    widget.onPressed();
+  }
+
+  void _onTapCancel() {
+    _controller.animateWith(
+      SpringSimulation(
+        const SpringDescription(mass: 1, stiffness: 300, damping: 20),
+        1.0,
+        0.0,
+        0.0,
+      ),
+    );
+  }
+
+  @override
+  void dispose() {
+    _controller.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      onTapDown: _onTapDown,
+      onTapUp: _onTapUp,
+      onTapCancel: _onTapCancel,
+      child: AnimatedBuilder(
+        animation: _controller,
+        builder: (context, child) {
+          final scale = 1.0 - (_controller.value * 0.1); // Scale down to 0.9
+          return Transform.scale(
+            scale: scale,
+            child: widget.child,
+          );
+        },
+      ),
     );
   }
 }
@@ -291,108 +386,96 @@ class WelcomeScreen extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       body: SafeArea(
-        child: Padding(
-          padding: const EdgeInsets.all(AppSpacing.screenPadding),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.stretch,
-            children: [
-              const SizedBox(height: AppSpacing.spacing48),
-
-              const Text(
-                'Welcome to',
-                style: AppTextStyles.bodyLarge,
-                textAlign: TextAlign.center,
-              ),
-              const SizedBox(height: AppSpacing.spacing8),
-
-              ShaderMask(
-                shaderCallback: (bounds) =>
-                    AppColors.primaryGradient.createShader(bounds),
-                child: const Text(
-                  'MindFlow',
-                  style: TextStyle(
-                    fontSize: 48,
-                    fontWeight: FontWeight.bold,
-                    color: Colors.white,
-                    height: 1.1,
+        child: SingleChildScrollView(
+          child: Padding(
+            padding: const EdgeInsets.all(AppSpacing.screenPadding),
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.stretch,
+              children: [
+                const SizedBox(height: AppSpacing.spacing48),
+                const Text(
+                  'Welcome to',
+                  style: AppTextStyles.bodyLarge,
+                  textAlign: TextAlign.center,
+                ),
+                const SizedBox(height: AppSpacing.spacing8),
+                ShaderMask(
+                  shaderCallback: (bounds) =>
+                      AppColors.primaryGradient.createShader(bounds),
+                  child: const Text(
+                    'MindFlow',
+                    style: TextStyle(
+                      fontSize: 48,
+                      fontWeight: FontWeight.bold,
+                      color: Colors.white,
+                      height: 1.1,
+                    ),
+                    textAlign: TextAlign.center,
+                  ),
+                ),
+                const SizedBox(height: AppSpacing.spacing12),
+                Text(
+                  'Your personal meditation\n& wellness companion',
+                  style: AppTextStyles.bodyMedium.copyWith(
+                    color: AppColors.neutralDark,
                   ),
                   textAlign: TextAlign.center,
                 ),
-              ),
-              const SizedBox(height: AppSpacing.spacing12),
-
-              Text(
-                'Your personal meditation\n& wellness companion',
-                style: AppTextStyles.bodyMedium.copyWith(
-                  color: AppColors.neutralDark,
+                const SizedBox(height: AppSpacing.spacing48),
+                const _FeatureCard(
+                  icon: Icons.self_improvement_rounded,
+                  iconGradient: AppColors.sageGradient,
+                  title: 'Mindful Meditation',
+                  description:
+                      'Guided sessions for stress, sleep, focus, and more.',
                 ),
-                textAlign: TextAlign.center,
-              ),
-
-              const SizedBox(height: AppSpacing.spacing48),
-
-              const _FeatureCard(
-                icon: Icons.self_improvement_rounded,
-                iconGradient: AppColors.sageGradient,
-                title: 'Mindful Meditation',
-                description:
-                    'Guided sessions for stress, sleep, focus, and more.',
-              ),
-
-              const SizedBox(height: AppSpacing.spacing16),
-
-              const _FeatureCard(
-                icon: Icons.nightlight_round,
-                iconColor: AppColors.accentBlueDark,
-                title: 'Better Sleep',
-                description:
-                    'Sleep stories, soundscapes, and wind-down exercises.',
-              ),
-
-              const SizedBox(height: AppSpacing.spacing16),
-
-              const _FeatureCard(
-                icon: Icons.trending_up_rounded,
-                iconColor: AppColors.primaryOrange,
-                title: 'Track Progress',
-                description:
-                    'Build healthy habits with streaks and daily goals.',
-              ),
-
-              const Spacer(),
-
-              AppButton(
-                text: 'Get Started',
-                icon: Icons.arrow_forward_rounded,
-                onPressed: () {
-                  Navigator.of(context).push(
-                    MaterialPageRoute(
-                      builder: (context) => ProfilingScreen(
-                        onComplete: (profile) {
-                          _navigateToMainApp(context, profile);
-                        },
+                const SizedBox(height: AppSpacing.spacing16),
+                const _FeatureCard(
+                  icon: Icons.nightlight_round,
+                  iconColor: AppColors.accentBlueDark,
+                  title: 'Better Sleep',
+                  description:
+                      'Sleep stories, soundscapes, and wind-down exercises.',
+                ),
+                const SizedBox(height: AppSpacing.spacing16),
+                const _FeatureCard(
+                  icon: Icons.trending_up_rounded,
+                  iconColor: AppColors.primaryOrange,
+                  title: 'Track Progress',
+                  description:
+                      'Build healthy habits with streaks and daily goals.',
+                ),
+                const SizedBox(height: AppSpacing.spacing32),
+                AppButton(
+                  text: 'Get Started',
+                  icon: Icons.arrow_forward_rounded,
+                  onPressed: () {
+                    Navigator.of(context).push(
+                      MaterialPageRoute(
+                        builder: (context) => ProfilingScreen(
+                          onComplete: (profile) {
+                            _navigateToMainApp(context, profile);
+                          },
+                        ),
                       ),
+                    );
+                  },
+                ),
+                const SizedBox(height: AppSpacing.spacing16),
+                TextButton(
+                  onPressed: () {
+                    _navigateToMainApp(context, NLPProfile.defaultProfile);
+                  },
+                  child: Text(
+                    'Skip to app',
+                    style: AppTextStyles.buttonSmall.copyWith(
+                      color: AppColors.neutralMedium,
                     ),
-                  );
-                },
-              ),
-
-              const SizedBox(height: AppSpacing.spacing16),
-
-              TextButton(
-                onPressed: () {
-                  _navigateToMainApp(context, NLPProfile.defaultProfile);
-                },
-                child: Text(
-                  'Skip to app',
-                  style: AppTextStyles.buttonSmall.copyWith(
-                    color: AppColors.neutralMedium,
                   ),
                 ),
-              ),
-
-              const SizedBox(height: AppSpacing.spacing24),
-            ],
+                const SizedBox(height: AppSpacing.spacing24),
+              ],
+            ),
           ),
         ),
       ),
@@ -451,9 +534,7 @@ class _FeatureCard extends StatelessWidget {
             height: 48,
             decoration: BoxDecoration(
               gradient: iconGradient,
-              color: iconGradient == null
-                  ? const Color(0x26F4A261)
-                  : null,
+              color: iconGradient == null ? const Color(0x26F4A261) : null,
               borderRadius: BorderRadius.circular(12),
             ),
             child: Icon(
@@ -464,9 +545,7 @@ class _FeatureCard extends StatelessWidget {
               size: 24,
             ),
           ),
-
           const SizedBox(width: AppSpacing.spacing16),
-
           Expanded(
             child: Column(
               crossAxisAlignment: CrossAxisAlignment.start,
