@@ -5,15 +5,16 @@ import 'package:http/http.dart' as http;
 
 import '../../../core/config/env_config.dart' as env;
 import '../domain/models/message.dart';
+import '../../coach/domain/models/coach.dart';
 import '../../onboarding/domain/models/nlp_profile.dart';
 import '../domain/services/nlp_prompt_builder.dart';
 
 /// Gemini Service for MindFlow AI Coach
 /// Handles communication with backend API proxy for Gemini
 /// Uses NLP Prompt Builder for profile-adaptive system prompts
-/// 
+///
 /// SECURITY: API key is stored server-side only, never exposed to client
-/// 
+///
 /// Configuration (per Technical Specification):
 /// - Model: gemini-2.0-flash (fast, cost-effective)
 /// - Temperature: 0.7 (creative but coherent)
@@ -32,6 +33,12 @@ class GeminiService {
   /// Current system prompt (set per user profile)
   String? _currentSystemPrompt;
 
+  /// Last known profile for prompt regeneration
+  NLPProfile? _lastProfile;
+
+  /// Active coach for persona injection
+  Coach? _activeCoach;
+
   /// User progress context for coaching
   int? _currentStreak;
   int? _totalSessions;
@@ -46,18 +53,33 @@ class GeminiService {
     debugPrint('GeminiService: Initialized with backend proxy');
   }
 
+  /// Set the active coach
+  void setActiveCoach(Coach coach) {
+    _activeCoach = coach;
+    debugPrint('GeminiService: Active coach set to ${coach.name}');
+    
+    // Regenerate prompt if we have a profile
+    if (_lastProfile != null) {
+      setUserProfile(_lastProfile!);
+    }
+  }
+
   /// Set the system prompt based on user's NLP profile
   /// Call this when user logs in or profile changes
   void setUserProfile(NLPProfile profile) {
+    _lastProfile = profile;
+    
     _currentSystemPrompt = NLPPromptBuilder.generateSystemPrompt(
       profile,
+      coach: _activeCoach,
       currentStreak: _currentStreak,
       totalSessions: _totalSessions,
       totalMinutes: _totalMinutes,
       activeGoal: _activeGoal,
       goalProgress: _goalProgress,
     );
-    debugPrint('GeminiService: System prompt updated for ${profile.displayName}');
+    debugPrint(
+        'GeminiService: System prompt updated for ${profile.displayName} (Coach: ${_activeCoach?.name ?? "Default"})');
   }
 
   /// Set user progress context for more personalized coaching
@@ -113,21 +135,46 @@ class GeminiService {
 
     try {
       final limitedHistory = _getLimitedHistory(chatHistory);
-      final historyJson = limitedHistory.where((msg) => msg.content.isNotEmpty).map((msg) {
-        return {
-          'isUser': msg.isUser,
-          'content': msg.content,
-        };
-      }).toList();
+
+      // Construct prompt with history manually for direct API
+      final fullPromptBuffer = StringBuffer();
+      fullPromptBuffer.writeln(_currentSystemPrompt ?? '');
+      fullPromptBuffer.writeln('\n--- CONVERSATION HISTORY ---');
+
+      for (final msg in limitedHistory) {
+        fullPromptBuffer
+            .writeln('${msg.isUser ? "USER" : "ASSISTANT"}: ${msg.content}');
+      }
+
+      fullPromptBuffer.writeln('USER: $userMessage');
+      fullPromptBuffer.writeln('ASSISTANT:'); // Priming
+
+      final apiKey = env.EnvConfig.geminiApiKey;
+      
+      if (apiKey.isEmpty) {
+        debugPrint('GeminiService: ERROR - API key is empty!');
+        yield* _getMockResponse(userMessage, profile);
+        return;
+      }
+      
+      final apiUrl =
+          'https://generativelanguage.googleapis.com/v1beta/models/gemini-1.5-flash:streamGenerateContent?key=$apiKey';
+
+      debugPrint('GeminiService: Sending direct request to Gemini API (Key length: ${apiKey.length})');
 
       final requestBody = json.encode({
-        'message': userMessage,
-        'history': historyJson,
-        'system_prompt': _currentSystemPrompt,
+        "contents": [
+          {
+            "parts": [
+              {"text": fullPromptBuffer.toString()}
+            ]
+          }
+        ],
+        "generationConfig": {
+          "temperature": 0.7,
+          "maxOutputTokens": 800,
+        }
       });
-
-      final apiUrl = '${env.apiBaseUrl}/api/chat';
-      debugPrint('GeminiService: Sending request to $apiUrl');
 
       final request = http.Request('POST', Uri.parse(apiUrl));
       request.headers['Content-Type'] = 'application/json';
@@ -139,42 +186,113 @@ class GeminiService {
 
         if (streamedResponse.statusCode != 200) {
           final errorBody = await streamedResponse.stream.bytesToString();
-          debugPrint('GeminiService: API error: $errorBody');
-          yield 'I\'m having a moment of reflection. Could you try again?';
+          debugPrint('GeminiService: API error ${streamedResponse.statusCode}: $errorBody');
+
+          if (streamedResponse.statusCode == 400) {
+            yield 'I\'m having trouble understanding. Could you try rephrasing? (Invalid request)';
+          } else if (streamedResponse.statusCode == 429) {
+            final retryMatch =
+                RegExp(r'retry in (\d+(\.\d+)?)s').firstMatch(errorBody);
+            final waitSeconds = retryMatch != null
+                ? double.tryParse(retryMatch.group(1) ?? '')?.ceil() ?? 30
+                : 30;
+            yield 'I need a moment to recharge. Please try again in $waitSeconds seconds. 🧘';
+          } else if (streamedResponse.statusCode == 403) {
+            yield 'I\'m not authorized to help right now. Please check the API configuration.';
+          } else {
+            yield 'I\'m having a moment of reflection. Could you try again? (Error: ${streamedResponse.statusCode})';
+          }
           return;
         }
 
         final stream = streamedResponse.stream.transform(utf8.decoder);
-        String buffer = '';
+        String fullResponse = '';
 
+        // Accumulate the full streamed response
         await for (final chunk in stream) {
-          buffer += chunk;
-          
-          while (buffer.contains('\n\n')) {
-            final endIndex = buffer.indexOf('\n\n');
-            final line = buffer.substring(0, endIndex);
-            buffer = buffer.substring(endIndex + 2);
+          fullResponse += chunk;
+        }
 
-            if (line.startsWith('data: ')) {
-              final data = line.substring(6);
-              
-              if (data == '[DONE]') {
-                return;
-              }
+        debugPrint(
+            'GeminiService: Received response length: ${fullResponse.length}');
 
-              try {
-                final parsed = json.decode(data);
-                if (parsed['text'] != null) {
-                  yield parsed['text'] as String;
-                } else if (parsed['error'] != null) {
-                  debugPrint('GeminiService: Stream error: ${parsed['error']}');
-                  yield 'I\'m having a moment of reflection. Could you try again?';
-                  return;
+        // Parse the accumulated JSON response
+        // The response is a JSON array: [{...}, {...}, ...]
+        // Each object contains candidates[0].content.parts[0].text
+        try {
+          // Try to parse as JSON and extract text properly
+          final dynamic jsonData = json.decode(fullResponse);
+          final textParts = <String>[];
+
+          if (jsonData is List) {
+            for (final item in jsonData) {
+              if (item is Map<String, dynamic>) {
+                final candidates = item['candidates'] as List<dynamic>?;
+                if (candidates != null && candidates.isNotEmpty) {
+                  final content =
+                      candidates[0]['content'] as Map<String, dynamic>?;
+                  if (content != null) {
+                    final parts = content['parts'] as List<dynamic>?;
+                    if (parts != null) {
+                      for (final part in parts) {
+                        if (part is Map<String, dynamic> &&
+                            part['text'] != null) {
+                          textParts.add(part['text'].toString());
+                        }
+                      }
+                    }
+                  }
                 }
-              } catch (e) {
-                debugPrint('GeminiService: Parse error: $e');
               }
             }
+          }
+
+          if (textParts.isNotEmpty) {
+            // Yield the combined text with word-by-word streaming effect
+            final combinedText = textParts.join('');
+            debugPrint('GeminiService: Successfully got response (${combinedText.length} chars)');
+            final words = combinedText.split(' ');
+            for (int i = 0; i < words.length; i++) {
+              yield words[i];
+              if (i < words.length - 1) yield ' ';
+              await Future.delayed(const Duration(milliseconds: 15));
+            }
+          } else {
+            debugPrint('GeminiService: No text parts found in response, using fallback');
+            yield* _getMockResponse(userMessage, profile);
+          }
+        } catch (parseError) {
+          debugPrint('GeminiService: JSON parse error: $parseError');
+          // Fallback: try regex extraction
+          final regex =
+              RegExp(r'"text"\s*:\s*"((?:[^"\\]|\\.)*)?"', multiLine: true);
+          final matches = regex.allMatches(fullResponse);
+          final extractedTexts = <String>[];
+
+          for (final match in matches) {
+            final text = match.group(1);
+            if (text != null && text.isNotEmpty) {
+              // Unescape JSON string
+              final unescaped = text
+                  .replaceAll(r'\n', '\n')
+                  .replaceAll(r'\"', '"')
+                  .replaceAll(r'\\', '\\')
+                  .replaceAll(r'\t', '\t');
+              extractedTexts.add(unescaped);
+            }
+          }
+
+          if (extractedTexts.isNotEmpty) {
+            final combinedText = extractedTexts.join('');
+            final words = combinedText.split(' ');
+            for (int i = 0; i < words.length; i++) {
+              yield words[i];
+              if (i < words.length - 1) yield ' ';
+              await Future.delayed(const Duration(milliseconds: 15));
+            }
+          } else {
+            debugPrint('GeminiService: Regex extraction also failed');
+            yield* _getMockResponse(userMessage, profile);
           }
         }
       } finally {
@@ -193,7 +311,7 @@ class GeminiService {
     NLPProfile? profile,
   }) async {
     final buffer = StringBuffer();
-    
+
     await for (final chunk in sendMessageStream(
       userMessage: userMessage,
       chatHistory: chatHistory,
@@ -201,7 +319,7 @@ class GeminiService {
     )) {
       buffer.write(chunk);
     }
-    
+
     return buffer.toString();
   }
 
@@ -209,9 +327,9 @@ class GeminiService {
   /// Streams crisis resources and support information
   Stream<String> _getCrisisResponse() async* {
     await Future.delayed(const Duration(milliseconds: 300));
-    
+
     final response = NLPPromptBuilder.generateCrisisResponse();
-    
+
     final words = response.split(' ');
     for (int i = 0; i < words.length; i++) {
       yield words[i];
@@ -223,13 +341,14 @@ class GeminiService {
   /// Get a mock response for development/demo
   /// Implements NLP-adaptive language patterns
   /// NOTE: Crisis detection is handled in sendMessageStream, not here
-  Stream<String> _getMockResponse(String userMessage, NLPProfile? profile) async* {
+  Stream<String> _getMockResponse(
+      String userMessage, NLPProfile? profile) async* {
     final effectiveProfile = profile ?? NLPProfile.defaultProfile;
-    
+
     await Future.delayed(const Duration(milliseconds: 500));
 
     final isCrisis = NLPPromptBuilder.containsCrisisIndicators(userMessage);
-    
+
     final response = _generateAdaptiveResponse(
       userMessage: userMessage,
       profile: effectiveProfile,
@@ -328,7 +447,8 @@ class GeminiService {
   String _getHumorPhrase(NLPProfile profile) {
     if (profile.motivation == 'toward' && profile.reference == 'internal') {
       return "You're basically a goal-crushing machine. The only question is which goal gets crushed next. 💪";
-    } else if (profile.motivation == 'away_from' && profile.reference == 'external') {
+    } else if (profile.motivation == 'away_from' &&
+        profile.reference == 'external') {
       return "We've all been the person who sets 47 alarms and still somehow oversleeps. You're in good company. 😅";
     } else {
       switch (profile.thinking) {
@@ -348,6 +468,5 @@ class GeminiService {
   }
 
   /// Dispose resources
-  void dispose() {
-  }
+  void dispose() {}
 }
